+++ src/Secure.SecurityDatabaseSync.BLL/Tasks/DefaultSyncTask.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Secure.SecurityDatabaseSync.BLL.Services
+namespace Secure.SecurityDatabaseSync.BLL.Tasks
 {
     /// <inheritdoc cref="ISyncTask"/>
     public class DefaultSyncTask : ISyncTask, IDisposable
